Added utils module and updated modules to use the new util functions
