+++ src/slack_watchman/utils.py
@@ -1,0 +1,44 @@
+import time
+import json
+import dataclasses
+from typing import List, Dict
+
+
+def convert_timestamp(timestamp: str or int) -> str or None:
+    """ Converts epoch timestamp into human-readable time
+
+    Args:
+        timestamp: epoch timestamp in seconds
+    Returns:
+        String time in the format YYYY-mm-dd hh:mm:ss
+    """
+
+    if timestamp:
+        if isinstance(timestamp, str):
+            timestamp = timestamp.split('.', 1)[0]
+
+        return time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(int(timestamp)))
+    else:
+        return None
+
+
+def deduplicate_dataclass(input_list: list) -> List[Dict]:
+    """ Removes duplicates where results are returned by multiple queries
+    Nested class handles JSON encoding for dataclass objects
+
+    Args:
+        input_list: List of dataclass objects
+    Returns:
+        List of JSON objects with duplicates removed
+    """
+
+    class EnhancedJSONEncoder(json.JSONEncoder):
+        def default(self, o):
+            if dataclasses.is_dataclass(o):
+                return dataclasses.asdict(o)
+            return super().default(o)
+
+    json_set = {json.dumps(dictionary, sort_keys=True, cls=EnhancedJSONEncoder) for dictionary in input_list}
+
+    deduped_list = [json.loads(t) for t in json_set]
+    return {match.get('watchman_id'): match for match in reversed(deduped_list)}.values()
